fix(ci): python-coverage-shards — bitbucket_client.py exceeded 1500-line guardrail
The #477 head-SHA resolution work pushed src/awf/common/bitbucket_client.py
to 1515 lines, tripping test_first_party_code_files_stay_under_line_limit
(MAX_FIRST_PARTY_FILE_LINES = 1500). Following the file's own established
decomposition pattern (errors + parsing already split into sibling modules),
move the pure `_PRContext` dataclass and `effective_merge_strategies` helper
into bitbucket_client_parsing.py and import them back. Pure relocation, no
behavior change; file is now 1475 lines.
